% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replisim_report.R
\name{replisim_report}
\alias{replisim_report}
\title{RepliSim Report (PDF)}
\usage{
replisim_report(template = "RepliSim", resetStyleFiles = FALSE, ...)
}
\arguments{
\item{template}{character; Pandoc template to use for rendering. Pass
\code{"RepliSim"} to use the default example template}

\item{resetStyleFiles}{logical; should the style files (logo, cover, defs.tex) be overwritten with the default files?}

\item{...}{further arguments passed to \code{\link[rmarkdown]{pdf_document}}}
}
\description{
This function can be used to render the RepliSim Report report. The design
and layout is loosly oriented at MS Word defaults and based on the INWTlab business report.
}
\details{
The function serves as wrapper to \code{\link[rmarkdown]{pdf_document}}
only steering the selection of the template.
}
